WIP: refactor: separate Doc business and display logic
