+++ whelktool/src/main/groovy/datatool/scripts/mergeworks/Doc.groovy
@@ -2,7 +2,6 @@ package datatool.scripts.mergeworks
 
 import se.kb.libris.Normalizers
 import whelk.Document
-import whelk.JsonLd
 import whelk.Whelk
 
 import static datatool.scripts.mergeworks.Util.asList
@@ -32,12 +31,14 @@ class Doc {
     Whelk whelk
     Document doc
     Map work
-    Map framed
+    
     List<String> titles
 
     //FIXME
     Document ogDoc
 
+    DisplayDoc display
+    
     Doc(Whelk whelk, Document doc) {
         this.whelk = whelk
         this.doc = doc
@@ -50,6 +51,14 @@ class Doc {
         }
 
         return work
+    }
+    
+    DisplayDoc getView() {
+        if (!display) {
+            display = new DisplayDoc(this)
+        }
+
+        return display
     }
 
     static Map getWork(Whelk whelk, Document d) {
@@ -88,22 +97,7 @@ class Doc {
     boolean hasGenericTitle() {
         Util.hasGenericTitle(getMainEntity()['hasTitle'])
     }
-
-    private static String displayTitle(Map thing) {
-        thing['hasTitle'].collect { it['@type'] + ": " + it['flatTitle'] }.join(', ')
-    }
-
-    String mainEntityDisplayTitle() {
-        displayTitle(['hasTitle': Util.flatTitles(getMainEntity()['hasTitle'])])
-    }
-
-    String link() {
-        String base = Document.getBASE_URI().toString()
-        String kat = "katalogisering/"
-        String id = doc.shortId
-        return base + kat + id
-    }
-
+    
     boolean isMonograph() {
         getMainEntity()['issuanceType'] == 'Monograph'
     }
@@ -131,142 +125,10 @@ class Doc {
         }
         pages ? pages.max() : -1
     }
+    
 
-    // TODO...
-    String getDisplayText(String field) {
-        if (field == 'contribution') {
-            return contributorStrings().join("<br>")
-        } else if (field == 'classification') {
-            return classificationStrings().join("<br>")
-        } else if (field == 'instance title') {
-            return isInstance() ? (getMainEntity()['hasTitle'] ?: '') : ''
-        } else if (field == 'work title') {
-            // To load hasTitle from linked work in instanceOf we can use getFramed()
-            // However we then need to handle that getFramed() loads linked instances in hasTitle.source
-            // Prefer getMainEntity() for now
-            return isInstance() ? (getMainEntity()['instanceOf']['hasTitle'] ?: '') : (getMainEntity()['hasTitle'] ?: '')
-        } else if (field == 'instance type') {
-            return isInstance() ? getMainEntity()['@type'] : ''
-        } else if (field == 'editionStatement') {
-            return getMainEntity()['editionStatement'] ?: ''
-        } else if (field == 'responsibilityStatement') {
-            return getMainEntity()['responsibilityStatement'] ?: ''
-        } else if (field == 'encodingLevel') {
-            return encodingLevel()
-        } else if (field == 'publication') {
-            return chipString(getMainEntity()['publication'] ?: [])
-        } else if (field == 'identifiedBy') {
-            return chipString(getMainEntity()['identifiedBy'] ?: [])
-        } else if (field == 'extent') {
-            return chipString(getMainEntity()['extent'] ?: [])
-        } else if (field == 'reproductionOf') {
-            return reproductionOfLink()
-        } else {
-            return chipString(getWork().getOrDefault(field, []))
-        }
-    }
-
-    protected String chipString(def thing) {
-        Util.chipString(thing, whelk)
-    }
-
-    String tooltip(String string, String tooltip) {
-        """<abbr title="${tooltip}">${string}</abbr>"""
-    }
-
-    private String reproductionOfLink() {
-        def shortId = Util.getPathSafe(getMainEntity(), ['reproductionOf', '@id'])
-                ?.tokenize("/#")
-                ?.dropRight(1)
-                ?.last() ?: ''
-
-        return "<a href=\"#$shortId\">$shortId</a>"
-    }
-
-    private List classificationStrings() {
-        List path = isInstance() ? ['instanceOf', 'classification'] : ['classification']
-        List<Map> classification = Util.getPathSafe(getFramed(), path, [])
-        classification.collect() { c ->
-            StringBuilder s = new StringBuilder()
-            s.append(flatMaybeLinked(c['inScheme'], ['code', 'version']).with { it.isEmpty() ? it : it + ': ' })
-            s.append(flatMaybeLinked(c, ['code']))
-            return s.toString()
-        }
-    }
-
-    private List contributorStrings() {
-        List path = isInstance() ? ['instanceOf', 'contribution'] : ['contribution']
-        List contribution = Util.getPathSafe(getFramed(), path, [])
-
-        return contribution.collect { Map c ->
-            contributionStr(c)
-        }
-    }
-
-    protected Map getFramed() {
-        if (!framed) {
-            if (isInstance()) {
-                framed = JsonLd.frame(doc.getThingIdentifiers().first(), whelk.loadEmbellished(doc.shortId).data)
-            } else {
-                Document copy = doc.clone()
-                whelk.embellish(copy)
-                framed = JsonLd.frame(doc.getThingIdentifiers().first(), copy.data)
-            }
-        }
-
-        return framed
-    }
-
-    private String contributionStr(Map contribution) {
-        StringBuilder s = new StringBuilder()
-
-        if (contribution['@type'] == 'PrimaryContribution') {
-            s.append('<b>')
-        }
-
-        s.append(flatMaybeLinked(contribution['role'], ['code', 'label']).with { it.isEmpty() ? it : it + ': ' })
-        s.append(flatMaybeLinked(contribution['agent'], ['givenName', 'familyName', 'lifeSpan', 'name']))
-
-        if (contribution['@type'] == 'PrimaryContribution') {
-            s.append('</b>')
-        }
-
-        return s.toString()
-    }
-
-    static String flatten(Object o, List order, String mapSeparator = ': ') {
-        if (o instanceof String) {
-            return o
-        }
-        if (o instanceof List) {
-            return o
-                    .collect { flatten(it, order) }
-                    .join(' || ')
-        }
-        if (o instanceof Map) {
-            return order
-                    .findResults { ((Map) o).get(it) }
-                    .collect { flatten(it, order) }
-                    .join(mapSeparator)
-        }
-
-        throw new RuntimeException(String.format("unexpected type: %s for %s", o.class.getName(), o))
-    }
-
-    private String flatMaybeLinked(Object thing, List order) {
-        if (!thing)
-            return ''
-
-        if (thing instanceof List) {
-            return thing.collect { flatMaybeLinked(it, order) }.join(' | ')
-        }
-        String s = flatten(thing, order, ', ')
-
-        thing['@id']
-                ? """<a href="${thing['@id']}">$s</a>"""
-                : s
-    }
-
+    
+    
     boolean isFiction() {
         isMarcFiction() || isSaogfFiction() || isSabFiction()
     }
@@ -284,12 +146,12 @@ class Doc {
     }
 
     boolean isSabFiction() {
-        classificationStrings().any { it.contains('kssb') && it.contains(': H') }
+        view.classificationStrings().any { it.contains('kssb') && it.contains(': H') }
     }
 
     boolean isNotFiction() {
         // A lot of fiction has marc/NotFictionNotFurtherSpecified but then classification is usually empty
-        isMarcNotFiction() && (!classificationStrings().isEmpty() && !isSabFiction())
+        isMarcNotFiction() && (!view.classificationStrings().isEmpty() && !isSabFiction())
     }
 
     boolean isText() {
@@ -305,7 +167,7 @@ class Doc {
     }
 
     boolean isSabDrama() {
-        classificationStrings().any { it.contains(': Hc.02') || it.contains(': Hce.02') }
+        view.classificationStrings().any { it.contains(': Hc.02') || it.contains(': Hce.02') }
     }
 
     boolean isGfDrama() {
